Print the elapsed time with a cast rather than %ld
print_time_taken() holds a millisecond count in a time_t and prints it
with %ld. That is only correct where time_t is long. On OpenBSD time_t
is __int64_t on every architecture, so on the 32 bit ones long is half
its width and the format is wrong: -Wformat rejects it, and where it
compiles it reads the wrong half of the value.

Cast at the call rather than changing the variable, so the diff stays to
the two lines that are wrong.

OpenBSD's ports tree has carried both lines as local patches, changing
the format to %lld without adding the cast. That is right for OpenBSD,
where time_t is always 64 bit, but wrong where time_t is long.

+++ examples/dump.c
@@ -119,7 +119,7 @@ static void print_time_taken(const struct timeval *start,
                              const struct timeval *stop) {
     time_t elapsed = (stop->tv_sec - start->tv_sec)*1000
                    + (stop->tv_usec - start->tv_usec)/1000;
-    fprintf(stderr, "time: %ld ms\n", elapsed);
+    fprintf(stderr, "time: %lld ms\n", (long long) elapsed);
 }
 
 int main(int argc, char **argv) {
